Factor out chunked_bulk_create
Preparing to use with Citation objects
